freelance-pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
google-generativeai>=0.8.0
schedule>=1.2.0
